main renaming work
