+++ CONTRIBUTING.md
@@ -2,32 +2,32 @@
 
 ## Fork and clone the repo
 
-Go to https://github.com/scikit-allel/scikit-allel-model and click the
+Go to https://github.com/scikit-allel/skallel-tensor and click the
 "Fork" button to fork the repository to your own GitHub account.
 
-Clone your fork of the scikit-allel-model repo to your local computer:
+Clone your fork of the skallel-tensor repo to your local computer:
 
 ```bash
-git clone git@github.com:username/scikit-allel-model.git
-cd scikit-allel-model
+git clone git@github.com:username/skallel-tensor.git
+cd skallel-tensor
 ```
 
 ...replacing "username" with your own GitHub username.
 
 The rest of this guide assumes your current working directory is your
-local clone of the scikit-allel-model repo.
+local clone of the skallel-tensor repo.
 
 ## Install development environment
 
 Make sure you have Python 3.7 installed on your local system somehow.
 
 Create a virtual environment for development work on the
-scikit-allel-model distribution. E.g., install
+skallel-tensor distribution. E.g., install
 [virtualenvwrapper](https://virtualenvwrapper.readthedocs.io/en/latest/),
 then run:
 
 ```bash
-mkvirtualenv --python=/usr/bin/python3.7 scikit-allel-model
+mkvirtualenv --python=/usr/bin/python3.7 skallel-tensor
 pip install -r requirements.txt
 pip install -e .
 ```
@@ -35,11 +35,11 @@ pip install -e .
 ...replacing "/usr/bin/python3.7" with whatever is the path to your
 Python 3.7 binary.
 
-Whenever you want to do development work on scikit-allel-model,
+Whenever you want to do development work on skallel-tensor,
 remember to activate the virtual environment, e.g.:
 
 ```bash
-workon scikit-allel-model
+workon skallel-tensor
 ```
 
 ## Run tests
