+++ README.md
@@ -1,4 +1,4 @@
-# scikit-allel-model
+# skallel-tensor
 
 Array functions and classes for genetic variation data.
 
@@ -7,7 +7,7 @@ repo.
 
 See [ROADMAP.md](ROADMAP.md) for design notes and the development roadmap.
 
-[![PyPI version](https://badge.fury.io/py/scikit-allel-model.svg)](https://badge.fury.io/py/scikit-allel-model)
-[![Build Status](https://travis-ci.org/scikit-allel/scikit-allel-model.svg?branch=master)](https://travis-ci.org/scikit-allel/scikit-allel-model)
-[![Build status](https://ci.appveyor.com/api/projects/status/8yld0akrlsl7oeot?svg=true)](https://ci.appveyor.com/project/alimanfoo/scikit-allel-model)
-[![Coverage Status](https://coveralls.io/repos/github/scikit-allel/scikit-allel-model/badge.svg?branch=master)](https://coveralls.io/github/scikit-allel/scikit-allel-model?branch=master)
+[![PyPI version](https://badge.fury.io/py/skallel-tensor.svg)](https://badge.fury.io/py/skallel-tensor)
+[![Build Status](https://travis-ci.org/scikit-allel/skallel-tensor.svg?branch=master)](https://travis-ci.org/scikit-allel/skallel-tensor)
+[![Build status](https://ci.appveyor.com/api/projects/status/8yld0akrlsl7oeot?svg=true)](https://ci.appveyor.com/project/alimanfoo/skallel-tensor)
+[![Coverage Status](https://coveralls.io/repos/github/scikit-allel/skallel-tensor/badge.svg?branch=master)](https://coveralls.io/github/scikit-allel/skallel-tensor?branch=master)
