+++ asv.conf.json
@@ -4,10 +4,10 @@
     "version": 1,
 
     // The name of the project being benchmarked
-    "project": "scikit-allel-model",
+    "project": "skallel-tensor",
 
     // The project's homepage
-    "project_url": "https://github.com/scikit-allel/scikit-allel-model",
+    "project_url": "https://github.com/scikit-allel/skallel-tensor",
 
     // The URL or local path of the source code repository for the
     // project being benchmarked
@@ -51,7 +51,7 @@
     //"install_timeout": 600,
 
     // the base URL to show a commit for the project.
-    "show_commit_url": "https://github.com/scikit-allel/scikit-allel-model/commit/",
+    "show_commit_url": "https://github.com/scikit-allel/skallel-tensor/commit/",
 
     // The Pythons you'd like to test against.  If not provided, defaults
     // to the current version of Python used to run `asv`.
