+++ requirements.txt
@@ -1,8 +1,11 @@
+# Pinned requirements used for CI and development environments.
+#
 # library requirements
 numba==0.44.1
 numpy==1.16.4
 dask[complete]==2.0.0
 pandas==0.24.2
+#
 # test requirements
 tox==3.13.1
 pytest==5.0.0
