+++ setup.py
@@ -1,18 +1,16 @@
-from setuptools import setup
+from setuptools import setup, find_packages
 
 setup(
-    name="scikit-allel-model",
-    description="Array functions and classes for genetic variation data.",
-    # N.B., cannot use find_packages() with native (implicit) namespace package
-    # https://packaging.python.org/guides/packaging-namespace-packages/#native-namespace-packages
-    packages=["skallel.model"],
+    name="skallel-tensor",
+    description="Tensor functions for genetic variation data.",
+    packages=find_packages('src'),
     package_dir={"": "src"},
     setup_requires=["setuptools>18.0", "setuptools-scm>1.5.4"],
     install_requires=["numpy", "numba", "dask>=2.0.0"],
     use_scm_version={
         "version_scheme": "guess-next-dev",
         "local_scheme": "dirty-tag",
-        "write_to": "src/skallel/model/version.py",
+        "write_to": "src/skallel_tensor/version.py",
     },
     classifiers=[
         "Development Status :: 3 - Alpha",
@@ -28,7 +26,7 @@ setup(
     ],
     maintainer="Alistair Miles",
     maintainer_email="alimanfoo@googlemail.com",
-    url="https://github.com/scikit-allel/scikit-allel-model",
+    url="https://github.com/scikit-allel/skallel-tensor",
     license="MIT",
     include_package_data=True,
     zip_safe=False,
