--- src/skallel/README.md
@@ -1,2 +1,0 @@
-N.B., this is a [namespace package](https://www.python.org/dev/peps/pep-0420/). **Do 
-not** add an `__init__.py` file.
