--- src/skallel/model/__init__.py
@@ -1,1 +1,0 @@
-from skallel.model.version import version as __version__  # noqa
