+++ src/skallel_tensor/__init__.py
@@ -1,0 +1,1 @@
+from .version import version as __version__  # noqa
