+++ src/skallel_tensor/methods_numpy.py
