+++ src/skallel_tensor/test/test_genotype_array.py
@@ -5,7 +5,7 @@ from numpy.testing import assert_array_equal
 import zarr
 
 
-from skallel.model.functions import (
+from skallel_tensor.functions import (
     genotype_array_check,
     genotype_array_is_called,
     genotype_array_is_missing,
