+++ tox.ini
@@ -1,15 +1,16 @@
 [tox]
-envlist = py{36,37}-{jit,nojit}
+envlist = py{36,37}-{jit,cov}
 
 [testenv]
 deps = -r requirements.txt
 usedevelop = true
 setenv =
-    py{36,37}-nojit: NUMBA_DISABLE_JIT=1
+    # disable JIT to get coverage
+    py{36,37}-cov: NUMBA_DISABLE_JIT=1
 commands =
     flake8 --max-line-length=88 src tests benchmarks setup.py
     # for coverage, run tests without numba jit compilation
-    py{36,37}-nojit: pytest -vv --cov=skallel --cov-fail-under=100 --ignore=src
-    py{36,37}-nojit: coverage report -m
+    py{36,37}-cov: pytest -vv --cov=skallel --cov-fail-under=100 --ignore=src
+    py{36,37}-cov: coverage report -m
     # also run tests with numba jit compilation
-    py{36,37}-jit: pytest -vv --ignore=src
+    py{36,37}-jit pytest -vv --ignore=src
